Re-arrange cli tests into separate query/tx files
